Extract dummy app into monorepo setup
