+++ ember-ace/addon-test-support/helpers/editor-interaction.ts
