+++ ember-ace/addon-test-support/helpers/poll-condition.ts
