+++ ember-ace/addon/components/ace-editor.hbs
