+++ ember-ace/addon/components/ace-editor.ts
