+++ ember-ace/app/components/ace-editor.js
