+++ ember-ace/index.js
