+++ ember-ace/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "ember-ace",
+  "version": "2.0.2",
+  "description": "An Ember addon for the Ace code editor",
+  "engines": {
+    "node": "14.* || >= 16"
+  },
+  "keywords": ["ember-addon"],
+  "repository": "dfreeman/ember-ace",
+  "license": "MIT",
+  "author": "Dan Freeman",
+  "dependencies": {
+    "ember-auto-import": "^2.4.0",
+    "ember-cli-babel": "^7.26.11",
+    "ember-cli-htmlbars": "^6.0.1",
+    "ember-cli-typescript": "^5.1.0",
+    "ember-modifier": "^3.2.7"
+  },
+  "peerDependencies": {
+    "ace-builds": "^1.8.0",
+    "ember-cli-page-object": "^1.17.9 || ^2.0.0-beta.0"
+  },
+  "peerDependenciesMeta": {
+    "ember-cli-page-object": {
+      "optional": true
+    }
+  },
+  "volta": {
+    "extends": "../package.json"
+  }
+}
