--- ember-cli-build.js
@@ -1,36 +1,0 @@
-'use strict';
-
-const EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
-
-module.exports = function (defaults) {
-  let app = new EmberAddon(defaults, {
-    autoImport: {
-      webpack: {
-        module: {
-          rules: [
-            {
-              resourceQuery: /resource/,
-              type: 'asset/resource',
-            },
-          ],
-        },
-      },
-    },
-  });
-
-  /*
-    This build file specifies the options for the dummy test app of this
-    addon, located in `/tests/dummy`
-    This build file does *not* influence how the addon or the app using it
-    behave. You most likely want to be modifying `./index.js` or app's build file
-  */
-
-  const { maybeEmbroider } = require('@embroider/test-setup');
-  return maybeEmbroider(app, {
-    skipBabel: [
-      {
-        package: 'qunit',
-      },
-    ],
-  });
-};
