+++ test-app-classic/.template-lintrc.js
