+++ test-app-classic/.watchmanconfig
