+++ test-app-classic/app/router.ts
@@ -1,5 +1,5 @@
 import EmberRouter from '@ember/routing/router';
-import config from 'dummy/config/environment';
+import config from 'test-app-classic/config/environment';
 
 export default class Router extends EmberRouter {
   location = config.locationType;
