+++ test-app-classic/app/templates/application.hbs
@@ -1,4 +1,4 @@
-{{page-title 'Dummy'}}
+{{page-title 'Test Classic App'}}
 
 <h2 id='title'>Welcome to Ember</h2>
