+++ test-app-classic/config/ember-try.js
