+++ test-app-classic/package.json
@@ -1,0 +1,82 @@
+{
+  "name": "test-app-classic",
+  "private": true,
+  "version": "0.0.0",
+  "scripts": {
+    "build": "ember build --environment=production",
+    "lint": "npm-run-all --aggregate-output --continue-on-error --parallel \"lint:!(fix)\"",
+    "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*:fix",
+    "lint:hbs": "ember-template-lint .",
+    "lint:hbs:fix": "ember-template-lint . --fix",
+    "lint:js": "eslint . --cache",
+    "lint:js:fix": "eslint . --fix",
+    "start": "ember serve",
+    "test": "npm-run-all lint test:*",
+    "test:ember": "ember test",
+    "test:ember-compatibility": "ember try:each",
+    "prepack": "ember ts:precompile",
+    "postpack": "ember ts:clean"
+  },
+  "devDependencies": {
+    "@ember/optional-features": "^2.0.0",
+    "@ember/test-helpers": "^2.8.1",
+    "@embroider/test-setup": "^1.2.0",
+    "@glimmer/component": "^1.1.2",
+    "@glimmer/tracking": "^1.1.2",
+    "@tsconfig/ember": "^1.0.1",
+    "@types/ember": "^4.0.0",
+    "@types/ember-qunit": "^5.0.0",
+    "@types/ember-resolver": "^5.0.11",
+    "@types/ember__destroyable": "^4.0.0",
+    "@types/jquery": "^3.5.14",
+    "@types/sinon": "^10.0.13",
+    "@typescript-eslint/eslint-plugin": "^5.30.7",
+    "@typescript-eslint/parser": "^5.30.7",
+    "ace-builds": "^1.8.1",
+    "babel-eslint": "^10.1.0",
+    "ember-auto-import": "^2.4.0",
+    "ember-ace": "*",
+    "ember-cli": "~4.2.0",
+    "ember-cli-babel": "^7.26.11",
+    "ember-cli-dependency-checker": "^3.2.0",
+    "ember-cli-htmlbars": "^6.0.1",
+    "ember-cli-inject-live-reload": "^2.1.0",
+    "ember-cli-page-object": "^1.17.9",
+    "ember-cli-release": "^1.0.0-beta.2",
+    "ember-cli-sri": "^2.1.1",
+    "ember-cli-terser": "^4.0.2",
+    "ember-cli-typescript": "^5.1.0",
+    "ember-destroyable-polyfill": "^2.0.3",
+    "ember-disable-prototype-extensions": "^1.1.3",
+    "ember-load-initializers": "^2.1.2",
+    "ember-page-title": "^7.0.0",
+    "ember-qunit": "^5.1.5",
+    "ember-resolver": "^8.0.3",
+    "ember-sinon": "^0.7.0",
+    "ember-source": "~3.28.0",
+    "ember-source-channel-url": "^3.0.0",
+    "ember-template-lint": "^4.2.0",
+    "ember-try": "^2.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.4.0",
+    "eslint-plugin-ember": "^10.5.9",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-qunit": "^7.2.0",
+    "loader.js": "^4.7.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.5.1",
+    "qunit": "^2.18.0",
+    "qunit-dom": "^2.0.0",
+    "typescript": "^4.7.4",
+    "webpack": "^5.69.1"
+  },
+  "license": "MIT",
+  "author": "Dan Freeman",
+  "ember": {
+    "edition": "octane"
+  },
+  "volta": {
+    "extends": "../package.json"
+  }
+}
