+++ test-app-classic/tests/test-helper.ts
@@ -1,5 +1,5 @@
-import Application from 'dummy/app';
-import config from 'dummy/config/environment';
+import Application from 'test-app-classic/app';
+import config from 'test-app-classic/config/environment';
 import * as QUnit from 'qunit';
 import { setApplication } from '@ember/test-helpers';
 import { setup } from 'qunit-dom';
